+++ R/collapse_rows.R
@@ -57,6 +57,10 @@ collapse_rows <- function(kable_input, columns = NULL,
                           col_names = TRUE,
                           longtable_clean_cut = TRUE) {
   kable_format <- attr(kable_input, "format")
+  if (kable_format %in% c("pipe", "markdown")) {
+    kable_input <- md_table_parser(kable_input)
+    kable_format <- attr(kable_input, "format")
+  }
   if (!kable_format %in% c("html", "latex")) {
     warning("Please specify format in kable. kableExtra can customize either ",
             "HTML or LaTeX outputs. See https://haozhu233.github.io/kableExtra/ ",
@@ -87,8 +91,7 @@ collapse_rows_html <- function(kable_input, columns, valign, target) {
   kable_xml <- kable_as_xml(kable_input)
   kable_tbody <- xml_tpart(kable_xml, "tbody")
 
-  kable_dt <- rvest::html_table(xml2::read_html(as.character(kable_input)))[[1]]
-  kable_dt <- as.data.frame(kable_dt)
+  kable_dt <- read_table_data_from_xml(kable_xml)
   if (is.null(columns)) {
     columns <- seq(1, ncol(kable_dt))
   }
@@ -96,11 +99,6 @@ collapse_rows_html <- function(kable_input, columns, valign, target) {
     if (!target %in% columns) {
       stop("target has to be within the range of columns")
     }
-  }
-  if (!is.null(kable_attrs$header_above)) {
-    kable_dt_col_names <- unlist(kable_dt[kable_attrs$header_above, ])
-    kable_dt <- kable_dt[-(1:kable_attrs$header_above),]
-    names(kable_dt) <- kable_dt_col_names
   }
   collapse_matrix <- collapse_row_matrix(kable_dt, columns, target = target)
 
@@ -208,9 +206,20 @@ collapse_rows_latex <- function(kable_input, columns, latex_hline, valign,
           new_kable_dt[i, columns[j]] <- ''
         }
       } else {
+        # We need to account for the cmidrules that
+        # are to the right of this column
+        num_rows <- collapse_matrix[i, j]
+        num_cols <- ncol(collapse_matrix) - j
+        if (num_rows && num_cols && valign != "\\[b\\]") {
+          vmove <- sum(rowSums(collapse_matrix[i - seq_len(num_rows-1), j + seq_len(num_cols), drop = FALSE]) > 0)
+          if (valign == "")
+            vmove <- 0.5*vmove
+        } else
+          vmove <- 0
         new_kable_dt[i, columns[j]] <- collapse_new_dt_item(
           kable_dt[i, columns[j]], collapse_matrix[i, j], column_width,
-          align = column_align, valign = valign
+          align = column_align, valign = valign,
+          vmove = vmove
         )
       }
     }
@@ -312,13 +321,15 @@ kable_dt_latex <- function(x, col_names) {
   data.frame(do.call(rbind, str_split(x, " & ")), stringsAsFactors = FALSE)
 }
 
-collapse_new_dt_item <- function(x, span, width = NULL, align, valign) {
+collapse_new_dt_item <- function(x, span, width = NULL, align, valign, vmove = 0) {
   if (span == 0) return("")
   if (span == 1) return(x)
   out <- paste0(
     "\\\\multirow", valign, "\\{", -span, "\\}\\{",
     ifelse(is.null(width), "\\*", width),
-    "\\}\\{",
+    "\\}",
+    if(vmove) paste0("[", vmove, "\\\\dimexpr\\\\aboverulesep+\\\\belowrulesep+\\\\cmidrulewidth]"),
+    "\\{",
     switch(align,
            "l" = "\\\\raggedright\\\\arraybackslash ",
            "c" = "\\\\centering\\\\arraybackslash ",